+++ multitenancy-spring-boot-3-integrationtests/pom.xml
@@ -36,7 +36,7 @@
     <packaging>jar</packaging>
 
     <properties>
-        <axon.version>4.10.0-SNAPSHOT</axon.version>
+        <axon.version>4.10.0</axon.version>
         <testcoontainers.version>1.20.0</testcoontainers.version>
 
         <jar-plugin.version>3.4.2</jar-plugin.version>
