+++ multitenancy/pom.xml
@@ -62,5 +62,9 @@
             <version>${mockito.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.axonframework</groupId>
+            <artifactId>axon-server-connector</artifactId>
+        </dependency>
     </dependencies>
 </project>
