+++ multitenancy/src/main/java/org/axonframework/extensions/multitenancy/components/eventstore/TenantPersistentStreamMessageSourceFactory.java
@@ -1,4 +1,19 @@
-package org.axonframework.extensions.multitenancy.autoconfig;
+/*
+ * Copyright (c) 2010-2024. Axon Framework
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.axonframework.extensions.multitenancy.components.eventstore;
 
 import io.axoniq.axonserver.connector.event.PersistentStreamProperties;
 import org.axonframework.axonserver.connector.AxonServerConfiguration;
@@ -16,7 +31,7 @@ import java.util.concurrent.ScheduledExecutorService;
  *
  * This interface is used to create a {@link PersistentStreamMessageSource} for a given tenant,
  * @author Stefan Dragisic
- * @since 4.10
+ * @since 4.10.0
  */
 @FunctionalInterface
 public interface TenantPersistentStreamMessageSourceFactory {
