+++ pom.xml
@@ -54,7 +54,7 @@
             ${project.basedir}/../coverage-report/target/site/jacoco-aggregate/jacoco.xml
         </sonar.coverage.jacoco.xmlReportPaths>
 
-        <axon.version>4.10.0-SNAPSHOT</axon.version>
+        <axon.version>4.10.0</axon.version>
 
         <spring.version>5.3.37</spring.version>
         <spring.boot.version>2.7.18</spring.boot.version>
